+++ src/rust/iced-x86/src/test_utils/mod.rs
@@ -70,6 +70,6 @@ pub(crate) fn get_default_ip(bitness: u32) -> u64 {
 pub(crate) fn create_decoder<'a>(bitness: u32, bytes: &'a [u8], options: u32) -> (Decoder<'a>, usize, bool) {
 	let mut decoder = Decoder::new(bitness, bytes, options);
 	decoder.set_ip(get_default_ip(bitness));
-	let len = cmp::min(IcedConstants::MAX_INSTRUCTION_LENGTH as usize, bytes.len());
+	let len = cmp::min(IcedConstants::MAX_INSTRUCTION_LENGTH, bytes.len());
 	(decoder, len, len < bytes.len())
 }
